Add more tests in Subscribable and integrate Subscribable in Events
