+++ htdocs/tests/js/unit/Events.spec.js
@@ -6,43 +6,8 @@ let chai = require('chai')
                 .use(require('sinon-chai'));
 let expect = chai.expect;
 let Events = require('./../../../resources/assets/js/models/Events');
+let HttpServiceMock = require('./../mocks/HttpServiceMock');
 
-class CallMock {
-
-    constructor (method, uri, params) {
-        this.method = method;
-        this.uri = uri;
-        this.params = params;
-        this.resolve = function(){};
-        this.reject = function(){};
-        this.done =function(callback) {
-            this.resolve = callback;
-            return this;
-        };
-        this.fail = function(callback) {
-            this.reject = callback;
-            return this;
-        };
-    }
-}
-
-class HttpServiceMock {
-
-    /**
-     * based on Jquery API https://api.jquery.com/jquery.get/
-     */
-
-    constructor () {
-        this.calls = [];
-        this.resolve = function(){};
-    }
-
-    get (uri, params) {
-        let callmock = new CallMock('GET', uri, params);
-        this.calls.push(callmock);
-        return callmock;
-    }
-}
 
 describe('ErgaTerminal', function() {
     let unit;
